- added macros that allow Albert to stay native even when using    function pointers, method pointers, or RAM hooks to those
Allright, I know that this commit looks insane and it probably is. I use all the tricks in the book (and a few tricks that will never be in the book) to write a hand full of macros which allow Albert code to call methods and functions via pointers and hooks in emulator RAM. Sounds boring, but this is quite a treat, because calling a method pointer to a method that already exists in Albert will not call the emulator at all. If the method is not yet implemented, Albert will simulate the register setup and call the emulator instead.

All this without changing any of the existing ported functions... . And of course these macros can be replaced with a single instruction when compiling Albert OS standalone.